Fixes Rakefile (mostly) so that rdoc and rake are up-to-date.
rdoc has some sort of deprecation warning that will pop up if you don't
'bundle exec rake' which is annoying. Not sure what to do about that
just yet, but it *is* a deprecation warning, so I needed to do
something.

I also put my email in and my name to reflect the fact that I will
maintain this fork.